+++ nilearn/datasets/func.py
@@ -1315,7 +1315,7 @@ def _load_mixed_gambles(zmap_imgs):
     X = X[mask, :].T
     tmp = np.zeros(list(mask.shape) + [len(X)])
     tmp[mask, :] = X.T
-    mask_img = nibabel.Nifti1Image(mask.astype(np.int), affine)
+    mask_img = nibabel.Nifti1Image(mask.astype(int), affine)
     X = nibabel.four_to_three(nibabel.Nifti1Image(tmp, affine))
     return X, y, mask_img
 
@@ -1982,8 +1982,8 @@ def fetch_development_fmri(n_subjects=None, reduce_confounds=True,
                            age_group='both'):
     """Fetch movie watching based brain development dataset (fMRI)
 
-    The data is downsampled to 4mm resolution for convenience. The origin of
-    the data is coming from OpenNeuro. See Notes below.
+    The data is downsampled to 4mm resolution for convenience with a repetition time (TR)
+    of 2 secs. The origin of the data is coming from OpenNeuro. See Notes below.
 
     Please cite [1]_ if you are using this dataset.
 
@@ -2225,7 +2225,7 @@ def fetch_language_localizer_demo_dataset(data_dir=None, verbose=1):
 
 
 def fetch_bids_langloc_dataset(data_dir=None, verbose=1):
-    """Download language localizer example bids dataset.
+    """Download language localizer example :term:`bids<BIDS>` dataset.
 
     Parameters
     ----------
@@ -2265,7 +2265,7 @@ def fetch_bids_langloc_dataset(data_dir=None, verbose=1):
 def fetch_openneuro_dataset_index(data_dir=None,
                                   dataset_version='ds000030_R1.0.4',
                                   verbose=1):
-    """Download a file with OpenNeuro BIDS dataset index.
+    """Download a file with OpenNeuro :term:`BIDS` dataset index.
 
     Downloading the index allows to explore the dataset directories
     to select specific files to download. The index is a sorted list of urls.
@@ -2382,7 +2382,7 @@ def select_from_index(urls, inclusion_filters=None, exclusion_filters=None,
 
 
 def patch_openneuro_dataset(file_list):
-    """Add symlinks for files not named according to latest BIDS conventions.
+    """Add symlinks for files not named according to latest :term:`BIDS` conventions.
     """
     rep = {'_T1w_brainmask': '_desc-brain_mask',
            '_T1w_preproc': '_desc-preproc_T1w',
@@ -2410,7 +2410,7 @@ def patch_openneuro_dataset(file_list):
 def fetch_openneuro_dataset(
     urls=None, data_dir=None, dataset_version='ds000030_R1.0.4',
     verbose=1):
-    """Download OpenNeuro BIDS dataset.
+    """Download OpenNeuro :term:`BIDS` dataset.
 
     Parameters
     ----------