+++ nilearn/plotting/img_plotting.py
@@ -401,7 +401,7 @@ class _MNI152Template(SpatialImage):
             anat_img = load_mni152_template()
             anat_img = reorder_img(anat_img)
             data = get_data(anat_img)
-            data = data.astype(np.float)
+            data = data.astype(np.float64)
             anat_mask = ndimage.morphology.binary_fill_holes(data > 0)
             data = np.ma.masked_array(data, np.logical_not(anat_mask))
             self._affine = anat_img.affine
@@ -1979,16 +1979,14 @@ def plot_carpet(img, mask_img=None, mask_labels=None,
 
     Notes
     -----
-    This figure was originally developed in [1]_.
+    This figure was originally developed in :footcite:`POWER2017150`.
 
     In cases of long acquisitions (>800 volumes), the data will be downsampled
     to have fewer than 800 volumes before being plotted.
 
     References
     ----------
-    .. [1] Power, J. D. (2017). A simple but useful way to assess fMRI scan
-            qualities. Neuroimage, 154, 150-158. doi:
-            https://doi.org/10.1016/j.neuroimage.2016.08.009
+    .. footbibliography::
 
     """
     img = _utils.check_niimg_4d(img, dtype='auto')