+++ nilearn/plotting/tests/test_surf_plotting.py
@@ -3,6 +3,7 @@ import numpy as np
 import nibabel
 import matplotlib.pyplot as plt
 import pytest
+import re
 import tempfile
 
 from nilearn.plotting.img_plotting import MNI152TEMPLATE
@@ -10,7 +11,9 @@ from nilearn.plotting.surf_plotting import (plot_surf, plot_surf_stat_map,
                                             plot_surf_roi, plot_img_on_surf,
                                             plot_surf_contours)
 from nilearn.datasets import fetch_surf_fsaverage
+from nilearn.surface import load_surf_mesh
 from nilearn.surface.testing_utils import generate_surf
+from numpy.testing import assert_array_equal
 
 
 def test_plot_surf():
@@ -35,6 +38,41 @@ def test_plot_surf():
     plot_surf(mesh, bg_map=bg, colorbar=True)
     plot_surf(mesh, bg_map=bg, colorbar=True, cbar_vmin=0,
               cbar_vmax=150, cbar_tick_format="%i")
+
+    # Plot with avg_method
+    ## Test all built-in methods and check
+    mapp = rng.standard_normal(size=mesh[0].shape[0])
+    mesh_ = load_surf_mesh(mesh)
+    coords, faces = mesh_[0], mesh_[1]
+
+    for method in ['mean', 'median', 'min', 'max']:
+        display = plot_surf(mesh, surf_map=mapp, avg_method=method)
+        if method == 'mean':
+            agg_faces = np.mean(mapp[faces], axis=1)
+        elif method == 'median':
+            agg_faces = np.median(mapp[faces], axis=1)
+        elif method == 'min':
+            agg_faces = np.min(mapp[faces], axis=1)
+        elif method == 'max':
+            agg_faces = np.max(mapp[faces], axis=1)
+        vmin = np.min(agg_faces)
+        vmax = np.max(agg_faces)
+        agg_faces -= vmin
+        agg_faces /= (vmax - vmin)
+        cmap = plt.cm.get_cmap(plt.rcParamsDefault['image.cmap'])
+        assert_array_equal(
+            cmap(agg_faces),
+            display._axstack.as_list()[0].collections[0]._facecolors
+        )
+
+    ## Try custom avg_method
+    def custom_avg_function(vertices):
+        return vertices[0] * vertices[1] * vertices[2]
+    plot_surf(
+        mesh,
+        surf_map=rng.standard_normal(size=mesh[0].shape[0]),
+        avg_method=custom_avg_function
+    )
 
     # Save execution time and memory
     plt.close()
@@ -71,6 +109,43 @@ def test_plot_surf_error():
             mesh, surf_map=rng.standard_normal(size=(mesh[0].shape[0], 2))
         )
 
+    with pytest.raises(
+        ValueError,
+        match=(
+            "Array computed with the custom "
+            "function from avg_method does "
+            "not have the correct shape"
+        )
+    ):
+        def custom_avg_function(vertices):
+            return [vertices[0] * vertices[1], vertices[2]]
+
+        plot_surf(mesh, surf_map=rng.standard_normal(size=mesh[0].shape[0]), avg_method=custom_avg_function)
+
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "avg_method should be either "
+            "['mean', 'median', 'max', 'min'] "
+            "or a custom function"
+        )
+    ):
+        custom_avg_function = dict()
+
+        plot_surf(mesh, surf_map=rng.standard_normal(size=mesh[0].shape[0]), avg_method=custom_avg_function)
+
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Array computed with the custom function "
+            "from avg_method should be an array of "
+            "numbers (int or float)"
+        )
+    ):
+        def custom_avg_function(vertices):
+            return "string"
+
+        plot_surf(mesh, surf_map=rng.standard_normal(size=mesh[0].shape[0]), avg_method=custom_avg_function)
 
 def test_plot_surf_stat_map():
     mesh = generate_surf()
@@ -221,6 +296,13 @@ def test_plot_surf_roi():
         plot_surf_roi(mesh, roi_map=roi_map, ax=plt.gca(), figure=None,
                       output_file=tmp_file.name, colorbar=True)
 
+    # Test nans handling
+    parcellation[::2] = np.nan
+    img = plot_surf_roi(mesh, roi_map=parcellation)
+    # Check that the resulting plot facecolors contain no transparent faces
+    # (last column equals zero) even though the texture contains nan values
+    assert(mesh[1].shape[0] ==
+           ((img._axstack.as_list()[0].collections[0]._facecolors[:, 3]) != 0).sum())
     # Save execution time and memory
     plt.close()