+++ doc/whats_new.rst
@@ -1,5 +1,62 @@
-0.7.X
+0.7.2.dev
+=========
+
+NEW
+---
+
+Fixes
+-----
+
+- Fix detrending and temporal filtering order for confounders
+  in :func:`nilearn.signal.clean`, so that these operations are applied
+  in the same order as for the signals, i.e., first detrending and
+  then temporal filtering (https://github.com/nilearn/nilearn/issues/2730).
+- Fix number of attributes returned by the
+  `nilearn.glm.first_level.FirstLevelModel._get_voxelwise_model_attribute` method in the first level model.
+  It used to return only the first attribute, and now returns as many attributes as design matrices.
+- Plotting functions that show a stack of slices from a 3D image (e.g. 
+  :func:`nilearn.plotting.plot_stat_map`) will now plot the slices in the user 
+  specified order, rather than automatically sorting into ascending order
+  (https://github.com/nilearn/nilearn/issues/1155). 
+
+Enhancements
+------------
+
+- :func:`nilearn.plotting.view_markers` now accepts an optional argument `marker_labels` to provide labels to each marker.
+- :func:`nilearn.plotting.plot_surf` now accepts new values for `avg_method` argument, such as `min`, `max`, or even a custom python function to compute the value displayed for each face of the plotted mesh.
+- :func:`nilearn.plotting.view_img_on_surf` can now optionally pass through
+  parameters to :func:`nilearn.surface.vol_to_surf` using the
+  `vol_to_surf_kwargs` argument. One application is better HTML visualization of
+  atlases.
+  (https://nilearn.github.io/auto_examples/01_plotting/plot_3d_map_to_surface_projection.html)
+
+.. _v0.7.1:
+
+0.7.1
 =====
+
+**Released March 2021**
+
+HIGHLIGHTS
+----------
+
+- New atlas fetcher
+  :func:`nilearn.datasets.fetch_atlas_difumo` to download *Dictionaries of Functional Modes*,
+  or “DiFuMo”, that can serve as atlases to extract functional signals with different
+  dimensionalities (64, 128, 256, 512, and 1024). These modes are optimized to represent well
+  raw BOLD timeseries, over a with range of experimental conditions.
+
+- :class:`nilearn.decoding.Decoder` and :class:`nilearn.decoding.DecoderRegressor`
+  is now implemented with random predictions to estimate a chance level.
+
+- The functions :func:`nilearn.plotting.plot_epi`, :func:`nilearn.plotting.plot_roi`,
+  :func:`nilearn.plotting.plot_stat_map`, :func:`nilearn.plotting.plot_prob_atlas`
+  is now implemented with new display mode Mosaic. That implies plotting 3D maps
+  in multiple columns and rows in a single axes.
+
+- :func:`nilearn.plotting.plot_carpet` now supports discrete atlases.
+  When an atlas is used, a colorbar is added to the figure,
+  optionally with labels corresponding to the different values in the atlas.
 
 NEW
 ---
@@ -131,7 +188,7 @@ NEW
   interface with the Nifti files on disk.
 - Plot events file
   Use :func:`nilearn.plotting.plot_event` to visualize events file.
-  The function accepts the BIDS events file read using `pandas`
+  The function accepts the :term:`BIDS` events file read using `pandas`
   utilities.
 - Plotting function :func:`nilearn.plotting.plot_roi` can now plot ROIs
   in contours with `view_type` argument.
@@ -517,7 +574,7 @@ NEW
   and `examples/03_connectivity/plot_canica_analysis.py` into an improved
   `examples/03_connectivity/plot_compare_decomposition.py`.
 
-- The Localizer dataset now follows the BIDS organization.
+- The Localizer dataset now follows the :term:`BIDS` organization.
 
 Changes
 -------